refactor(tui): extract skipEscapeSequence, move stripAnsi to serialize
